Use separate external pillar for solutions Fixes: #956
