+++ salt/metalk8s/salt/master/files/master-99-metalk8s.conf.j2
@@ -12,6 +12,7 @@ ext_pillar:
   - metalk8s_endpoints: /etc/kubernetes/admin.conf
   - metalk8s_nodes: /etc/kubernetes/admin.conf
   - metalk8s_private: {}
+  - metalk8s_solutions: {}
 
 roster_defaults:
   minion_opts:
